Move doc strings from `cli::Command` to commands themselves
Fold `cli` module back into `main`.
