+++ pgdo-cli/src/command/shell.rs
@@ -7,6 +7,8 @@ use crate::{
     runner::{self, Runner},
 };
 
+/// Start a psql shell, creating and starting the cluster as necessary
+/// (DEFAULT).
 #[derive(clap::Args)]
 #[clap(next_help_heading = Some("Options for shell"))]
 pub struct Args {
